+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Add `PodCreateOptsBuilder::utsns`
 * Add `PodCreateOptsBuilder::volumes`
 * Add `PodCreateOptsBuilder::idmappings`
+* Add `PodCreateOptsBuilder::exit_policy`
 
 # 0.7.0
 * Make futures returned by methods Send + Sync